refactor: add .js file extensions for imports/exports
