+++ .babelrc
@@ -7,5 +7,8 @@
         "modules": "amd"
       }
     ]
+  ],
+  "plugins": [
+    "./tools/babel-plugin-import-path-requirejs.js"
   ]
 }
