+++ src/core/base-runner.js
@@ -1,13 +1,13 @@
 // Module core/base-runner
 // The module in charge of running the whole processing pipeline.
-import "./include-config";
-import "./override-configuration";
-import "./respec-ready";
-import "./jquery-enhanced"; // for backward compatibility
-import { done as postProcessDone } from "./post-process";
-import { done as preProcessDone } from "./pre-process";
-import { pub } from "./pubsubhub";
-import { removeReSpec } from "./utils";
+import "./include-config.js";
+import "./override-configuration.js";
+import "./respec-ready.js";
+import "./jquery-enhanced.js"; // for backward compatibility
+import { done as postProcessDone } from "./post-process.js";
+import { done as preProcessDone } from "./pre-process.js";
+import { pub } from "./pubsubhub.js";
+import { removeReSpec } from "./utils.js";
 
 export const name = "core/base-runner";
 const canMeasure = performance.mark && performance.measure;
