+++ src/core/biblio-db.js
@@ -8,7 +8,7 @@
  *
  */
 /* globals IDBKeyRange, DOMException */
-import { pub } from "./pubsubhub";
+import { pub } from "./pubsubhub.js";
 export const name = "core/biblio-db";
 
 const ALLOWED_TYPES = new Set(["alias", "reference"]);
