+++ src/core/biblio.js
@@ -5,13 +5,13 @@
 
 /* jshint jquery: true */
 /* globals console */
-import { biblioDB } from "./biblio-db";
-import { createResourceHint } from "./utils";
-import { pub } from "./pubsubhub";
+import { biblioDB } from "./biblio-db.js";
+import { createResourceHint } from "./utils.js";
+import { pub } from "./pubsubhub.js";
 export const biblio = {};
 
 // for backward compatibity
-export { wireReference, stringifyReference } from "./render-biblio";
+export { wireReference, stringifyReference } from "./render-biblio.js";
 
 export const name = "core/biblio";
