+++ src/core/data-include.js
@@ -9,8 +9,8 @@
 //  This module only really works when you are in an HTTP context, and will most likely
 //  fail if you are editing your documents on your local drive. That is due to security
 //  restrictions in the browser.
-import { pub } from "./pubsubhub";
-import { runTransforms } from "./utils";
+import { pub } from "./pubsubhub.js";
+import { runTransforms } from "./utils.js";
 
 export const name = "core/data-include";
