+++ src/core/data-tests.js
@@ -9,10 +9,10 @@
  *
  * Docs: https://github.com/w3c/respec/wiki/data-tests
  */
-import { lang as defaultLang } from "./l10n";
+import { lang as defaultLang } from "./l10n.js";
 import hyperHTML from "hyperhtml";
-import { pub } from "./pubsubhub";
-import { showInlineWarning } from "./utils";
+import { pub } from "./pubsubhub.js";
+import { showInlineWarning } from "./utils.js";
 const l10n = {
   en: {
     missing_test_suite_uri:
