+++ src/core/data-transform.js
@@ -11,7 +11,7 @@
 //  know what you are doing, you should be using a dedicated module instead. This feature
 //  is not actively supported and support for it may be dropped. It is not accounted for
 //  in the test suite, and therefore could easily break.
-import { runTransforms } from "./utils";
+import { runTransforms } from "./utils.js";
 
 export const name = "core/data-transform";
