+++ src/core/defaults.js
@@ -2,13 +2,13 @@
  * Sets the core defaults
  */
 export const name = "core/defaults";
-import { rule as checkCharset } from "./linter-rules/check-charset";
-import { rule as checkInternalSlots } from "./linter-rules/check-internal-slots";
-import { rule as checkPunctuation } from "./linter-rules/check-punctuation";
-import linter from "./linter";
-import { rule as localRefsExist } from "./linter-rules/local-refs-exist";
-import { rule as noHeadinglessSectionsRule } from "./linter-rules/no-headingless-sections";
-import { rule as noHttpPropsRule } from "./linter-rules/no-http-props";
+import { rule as checkCharset } from "./linter-rules/check-charset.js";
+import { rule as checkInternalSlots } from "./linter-rules/check-internal-slots.js";
+import { rule as checkPunctuation } from "./linter-rules/check-punctuation.js";
+import linter from "./linter.js";
+import { rule as localRefsExist } from "./linter-rules/local-refs-exist.js";
+import { rule as noHeadinglessSectionsRule } from "./linter-rules/no-headingless-sections.js";
+import { rule as noHttpPropsRule } from "./linter-rules/no-http-props.js";
 
 linter.register(
   noHttpPropsRule,
