+++ src/core/exporter.js
@@ -6,10 +6,10 @@
  * That is, elements that have a "removeOnSave" css class.
  */
 
-import { expose } from "./expose-modules";
+import { expose } from "./expose-modules.js";
 import hyperHTML from "hyperhtml";
-import { pub } from "./pubsubhub";
-import { removeReSpec } from "./utils";
+import { pub } from "./pubsubhub.js";
+import { removeReSpec } from "./utils.js";
 
 const mimeTypes = new Map([["text/html", "html"], ["application/xml", "xml"]]);
