+++ src/core/fix-headers.js
@@ -2,7 +2,7 @@
 // Make sure that all h1-h6 headers (that are first direct children of sections) are actually
 // numbered at the right depth level. This makes it possible to just use any of them (conventionally
 // h2) with the knowledge that the proper depth level will be used
-import { renameElement } from "./utils";
+import { renameElement } from "./utils.js";
 
 export const name = "core/fix-headers";
