+++ src/core/github.js
@@ -4,7 +4,7 @@
  * @see https://github.com/w3c/respec/wiki/github
  */
 
-import { pub } from "./pubsubhub";
+import { pub } from "./pubsubhub.js";
 
 export const name = "core/github";
