+++ src/core/include-config.js
@@ -1,6 +1,6 @@
 // Module core/include-config
 // Inject's the document's configuration into the head as JSON.
-import { sub } from "./pubsubhub";
+import { sub } from "./pubsubhub.js";
 export const name = "core/include-config";
 
 const userConfig = {};
