+++ src/core/pubsubhub.js
@@ -1,4 +1,4 @@
-import { expose } from "./expose-modules";
+import { expose } from "./expose-modules.js";
 
 /**
  * Module core/pubsubhub
