+++ src/core/structure.js
@@ -9,7 +9,7 @@
 //  - lang: can change the generated text (supported: en, fr)
 //  - maxTocLevel: only generate a TOC so many levels deep
 
-import { addId, children, parents, renameElement } from "./utils";
+import { addId, children, parents, renameElement } from "./utils.js";
 import { lang as defaultLang } from "../core/l10n";
 import hyperHTML from "hyperhtml";
